Rename ApprovedBlockReceivedHandler to Running
