+++ casper/src/test/scala/coop/rchain/casper/engine/InitializingSpec.scala
@@ -14,7 +14,6 @@ import coop.rchain.casper.engine._, EngineCell._
 import coop.rchain.casper.helper.{BlockDagStorageTestFixture, NoOpsCasperEffect}
 import coop.rchain.casper.protocol.{NoApprovedBlockAvailable, _}
 import coop.rchain.casper.util.TestTime
-import coop.rchain.casper.util.comm.CasperPacketHandlerSpec._
 import coop.rchain.casper.util.rholang.RuntimeManager
 import coop.rchain.catscontrib.ApplicativeError_
 import coop.rchain.catscontrib.TaskContrib._
@@ -39,7 +38,7 @@ import scala.concurrent.duration._
 
 class InitializingSpec extends WordSpec {
   "Initializing state" should {
-    "make a transition to ApprovedBlockReceivedHandler once ApprovedBlock has been received" in {
+    "make a transition to Running once ApprovedBlock has been received" in {
       import monix.execution.Scheduler.Implicits.global
       val fixture = Setup()
       import fixture._
@@ -84,7 +83,7 @@ class InitializingSpec extends WordSpec {
         _               = assert(blockO.isDefined)
         _               = assert(blockO.contains(genesis))
         handlerInternal <- EngineCell[Task].read
-        _               = assert(handlerInternal.isInstanceOf[ApprovedBlockReceivedHandler[Task]])
+        _               = assert(handlerInternal.isInstanceOf[Running[Task]])
         _ = assert(
           transportLayer.requests.head.msg == packet(
             local,
