+++ casper/src/test/scala/coop/rchain/casper/helper/HashSetCasperTestNode.scala
@@ -93,7 +93,7 @@ class HashSetCasperTestNode[F[_]](
 
   implicit val multiparentCasperRef = MultiParentCasperRef.unsafe[F](Some(casperEff))
 
-  val engine                             = new ApprovedBlockReceivedHandler(casperEff, approvedBlock)
+  val engine                             = new Running(casperEff, approvedBlock)
   implicit val engineCell: EngineCell[F] = Cell.unsafe[F, Engine[F]](engine)
   val casperPacketHandler                = new CasperPacketHandler[F]
   implicit val packetHandlerEff = PacketHandler.pf[F](
